refactor: [development/AppRouter]-new pages and management LS
