+++ src/components/Navbar/Navbar.tsx
@@ -8,7 +8,7 @@ const eventListener = 'keydown'
 
 const Navbar = () => {
   const textLinks: AnchorLink[] = [
-    { type: "anchor", label: "Temporizador", href: "/", icon: "empty" },
+    { type: "anchor", label: "Temporizador", href: "/timer", icon: "empty" },
     { type: "anchor", label: "Hoy", href: "/today", icon: "empty" },
     { type: "anchor", label: "Tareas", href: "/task", icon: "empty" },
     { type: "anchor", label: "Registros", href: "/entries", icon: "empty" },
@@ -111,16 +111,30 @@ const Navbar = () => {
             }
           </section>
 
-          <Modal title="Configuraciones" id="modalSettings" className="absolute top-20 p-3 z-50 w-auto bg-background-light rounded-lg border-2 border-transparent hover:border-2 hover:outline-0 hover:border-hover-light transition ease-in-out duration-300">
-            <button onClick={() => handleOpenModal('modalPreferences')}>Administrar Preferencias</button>
-            <button onClick={() => console.log('Estado')}>Administrar Estados</button>
-            <button onClick={() => console.log('Datos')}>Gestión de Datos</button>
+          <Modal title="" id="modalSettings" className="absolute top-20 py-3 z-50 w-auto bg-background-light rounded-lg border-2 border-transparent hover:border-2 hover:outline-0 hover:border-hover-light transition ease-in-out duration-300">
+            <div className="flex flex-col items-start pt-2">
+              <div className="hover:bg-text-degraded/15 w-full py-2 pr-6 pl-2">
+                <button onClick={() => handleOpenModal('modalPreferences')}>Administrar Preferencias</button>
+              </div>
+              <div className="hover:bg-text-degraded/15 w-full py-2 pr-6 pl-2">
+                <button onClick={() => console.log('Estado')}>Administrar Estados</button>
+              </div>
+              <div className="hover:bg-text-degraded/15 w-full py-2 pr-6 pl-2">
+                <button onClick={() => console.log('Datos')}>Gestión de Datos</button>
+              </div>
+            </div>
           </Modal>
 
-          <Modal title="Soporte" id="modalSupport" className="absolute top-20 p-3 z-50 w-auto bg-background-light rounded-lg border-2 border-transparent hover:border-2 hover:outline-0 hover:border-hover-light transition ease-in-out duration-300">
-            <h2>¿Necesita ayuda?</h2>
-            <button onClick={() => console.log('FQA')}>Preguntas Frecuentes</button>
-            <button onClick={() => console.log('Contacto')}>Contactanos</button>
+          <Modal title="Soporte" id="modalSupport" className="absolute top-20 py-3 z-50 w-auto bg-background-light rounded-lg border-2 border-transparent hover:border-2 hover:outline-0 hover:border-hover-light transition-all ease-in-out duration-300">
+            <div className="flex flex-col items-start">
+              <h2 className="text-lg font-medium py-2 pr-6 pl-2 text-text-light dark:text-text-dark">¿Necesita ayuda?</h2>
+              <div className="hover:bg-text-degraded/15 w-full py-2 pr-6 pl-2">
+                <button onClick={() => console.log('FQA')}>Preguntas Frecuentes</button>
+              </div>
+              <div className="hover:bg-text-degraded/15 w-full py-2 pr-6 pl-2">
+                <button onClick={() => console.log('Contacto')}>Contactanos</button>
+              </div>
+            </div>
           </Modal>
 
           <Modal title="Preferencias" id="modalPreferences" className="absolute top-20 p-3 mx-3 z-50 w-auto bg-background-light rounded-lg border-2 border-transparent hover:border-2 hover:outline-0 hover:border-hover-light transition ease-in-out duration-300">
@@ -152,7 +166,7 @@ const Navbar = () => {
             ))
           }
           <div className="pt-4 pr-6 flex justify-end items-center border-t dark:border-text-dark/5 border-text-light/5">
-            <ThemeToggle id="scheme" parentMethod={handleTheme} watch="theme" iconOn="sun" iconOff="moon" />
+            <ThemeToggle id="scheme" description="Tema del Sistema" parentMethod={handleTheme} watch="theme" iconOn="sun" iconOff="moon" />
           </div>
         </div>
       </nav>
