+++ src/pages/ManagementPage.tsx
@@ -1,0 +1,7 @@
+const ManagementPage = () => {
+  return (
+    <div>Management Page</div>
+  )
+}
+
+export { ManagementPage }
