+++ src/pages/Task/StatusTaskPage.tsx
@@ -1,0 +1,10 @@
+const StatusTaskPage = () => {
+
+  return (
+    <div>
+      StatusTaskPage
+    </div>
+  )
+}
+
+export { StatusTaskPage }
