+++ src/pages/Task/index.ts
@@ -1,0 +1,2 @@
+export * from './StatusTaskPage';
+export * from './TaskPage';
