+++ src/pages/index.ts
@@ -1,8 +1,10 @@
 export * from './EntryPage';
+export * from './ManagementPage';
 export * from './NotFound/NotFound';
 export * from './NotFound/RoutesWithNotFound';
 export * from './SupportPage';
 export * from './TagPage';
-export * from './TaskPage';
+export * from './Task/StatusTaskPage';
+export * from './Task/TaskPage';
 export * from './TimerPage';
 export * from './TodayPage';
